reorganized source tree - making it little cleaner
