+++ include/utils.hpp
@@ -12,7 +12,8 @@ namespace kyber_utils {
 // Generates N -many random values of type T | N >= 0
 template<typename T>
 static inline void
-random_data(T* const data, const size_t len) requires(std::is_unsigned_v<T>)
+random_data(T* const data, const size_t len)
+  requires(std::is_unsigned_v<T>)
 {
   std::random_device rd;
   std::mt19937_64 gen(rd());
